cm 1221

+++ app/Http/Controllers/Admin/DX/ContentController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Admin\DX;
 
 use App\Models\DX\Content;
+use App\Models\DX\Quiz;
+use App\Models\DX\QuizAnswer;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
@@ -10,6 +12,16 @@ class ContentController extends Controller
 {
     public function index(){
 
+    }
+
+    public function show($id){
+        $content = Content::find($id);
+        $list = Quiz::where('content_id',$id)->get();
+        foreach ($list as $quiz){
+            $quiz->allAnswer = QuizAnswer::where('quiz_id',$quiz->id)->get();
+        }
+//        dd($list);
+        return view('Admin.DX.Course.contentShow',compact('list','content'));
     }
 
     public function create(Request $request){
@@ -31,7 +43,7 @@ class ContentController extends Controller
         $credentials = $this->validate($request,$verif);
 //        dd($credentials);
         if (Content::create($credentials)){
-            return redirect('admin/course/'.$credentials['course_id'])->with('success', config('hint.add_success'));
+            return redirect('admin/jbdx/course/'.$credentials['course_id'])->with('success', config('hint.add_success'));
         }else{
             return back()->with('hint',config('hint.add_failure'));
         }
@@ -57,7 +69,7 @@ class ContentController extends Controller
         $credentials = $this->validate($request,$verif);
 //        dd($credentials);
         if (Content::find($id)->update($credentials)){
-            return redirect('admin/course/'.$credentials['course_id'])->with('success',config('hint.mod_success'));
+            return redirect('admin/jbdx/course/'.$credentials['course_id'])->with('success',config('hint.mod_success'));
         }else{
             return back()->with('hint',config('hint.mod_failure'));
         }
@@ -68,6 +80,10 @@ class ContentController extends Controller
         $Content = Content::find($id);
         if (!$Content){
             return back() -> with('hint',config('hint.data_exist'));
+        }
+        $quiz = Quiz::where('content_id',$id)->get()->toArray();
+        if ($quiz){
+            return back()->with('hint',config('hint.del_failure_exist'));
         }
         if (Content::destroy($id)){
             return back() -> with('success',config('hint.del_success'));
